attempt to get balance widget showing - kyra
I was following a video that puts a widget on the screen to show the distance travelled, but they did it on the third person character BP, which inherits from 'Character'. I attempted to cast it as a character for this part only but it got really mad then so im not sure what to do. It doesn't crash or anything, the 0 on the screen just never updates.

the error message:
'VRPawn' does not inherit from 'Character' ( Cast To VRPawn  would always fail).