Added level change - Kyra
Under content, there's a block that is used as the collision detector which brings the player to my level for now. It's just a white box outside the front door for now. I also edited the landscape material under the landscape item for my level, to make the change a bit more obvious. Only thing is that it requires we do the ctrl i,m, and tab to get it back going again.